Do not install "futures" in case of Python 3
Since futures 3.2 [1], the package enforces to be installed only in Python 2
environments because it's basically a backport of Python 3's built-in
package. So in order to support both Python 2 and Python 3 environments, we
have to skip installation of futures package in case of Python 3.

It might look natural to use environment markers [2] to achieve this goal but
they are new and were introduced in setuptools in mid of 2017. FWIW,
setuptools on both Ubuntu Trusty and Ubuntu Xenial do not support them and
batch scoring script may be installed in pretty outdated envs. So let's do it
old-fashioned way by adding condition in setup.py.

[1] https://github.com/agronholm/pythonfutures/commit/d0393ad626d25622927bb0ed47d35ddb2f6cd321
[2] https://www.python.org/dev/peps/pep-0508/#environment-markers

Jira: API-4184