Build different wheels for Python 2 and Python 3
Since recently we have different set of dependencies for Python 2 and
Python 3 environments, hence we can't have one wheel because it may have
only one set of dependencies and no way of conditional dependencies.

So this commit removes "universal" wheels, which effectively means we
will have two wheels: one for python2, and one for python3.