Compile fresh OpenSSL inside dockerized env
Batch Scoring script uses Docker container based on manylinux [1] to
produce pyinstaller artifacts (i.e. all-in-one binaries). It turns out
that currently PyPI uses TLS 1.2, and this security protocol is not
supported by OpenSSL distributed inside manylinux container. So when we
build our own CPython interpreter (for the reasons mentioned in
corresponding Dockerfile) and link it against default OpenSSL, the
produced interpreter does not not support TLS 1.2. Therefore, when we
try to run "pip" and other tools that deal with PyPI, they are unable to
establish connection and fail.

In order to workaround this issue, let's compile our own fresh OpenSSL
library and link the produced interpreter against it.

Alternatively, it seems like there's no reason to use manylinux as base
image at all because we don't have dependencies that link against system
libraries. However, such change may be potentially destructive and
should be considered and implemented separately. After all, this is a
fix to unblock batch scoring releases.

[1] https://github.com/pypa/manylinux